Add more arguments

+++ tester.py
@@ -7,24 +7,38 @@ import argparse
 
 parser = argparse.ArgumentParser()
 
-parser.add_argument('--dir', default='.')
-parser.add_argument('--exec', default='a.out')
-parser.add_argument('--src', default='main.cpp')
-parser.add_argument('--testdir', default='tests')
+parser.add_argument('-d', '--dir', default='.')
+parser.add_argument('-o', '--exec', default='a.out')
+parser.add_argument('-i', '--src', default='main.cpp')
+parser.add_argument('-t', '--testdir', default='tests')
+parser.add_argument('-C', '--compiler', default='g++')
+parser.add_argument('-f', '--flags', default='-std=gnu++17 -O2 -static')
+parser.add_argument('-s', '--silent', default=False, action=argparse.BooleanOptionalAction)
+parser.add_argument('--compile', default=True, action=argparse.BooleanOptionalAction)
+parser.add_argument('-w', '--warnings', default=False, action=argparse.BooleanOptionalAction)
 
 args = vars(parser.parse_args())
 
 exec_path = os.path.join(args['dir'], args['exec'])
-compile_command = f"g++ {args['src']} -o {args['exec']}"
+warnings = '-Wall -Wextra -Wshadow' if args['warnings'] else '-w'
+compile_command = f"{args['compiler']} {args['flags']} {warnings} -o {args['exec']} {args['src']}"
 
 cat_cmd = 'type' if os.name == 'nt' else 'cat'
 
-if os.path.isfile(exec_path):
-    os.remove(exec_path)
+if args['compile']:
+    if os.path.isfile(exec_path):
+        os.remove(exec_path)
 
-os.popen(compile_command).read()
+    print(f'Compiling with: {compile_command}\n')
 
-if not os.path.isfile(exec_path):
+    os.popen(compile_command).read()
+
+    print()
+
+    if not os.path.isfile(exec_path):
+        exit(1)
+elif not os.path.isfile(exec_path):
+    print('Cant find executable file')
     exit(1)
 
 passed = 0
@@ -46,9 +60,14 @@ for in_name, out_name in get_test_names(args['testdir']):
             else:
                 failed += 1
                 print(f'Test {in_name} FAILED')
-                print('Expected:')
-                print(expected_output)
-                print('Got:')
+
+                if args['silent']:
+                    print('Result:')
+                else:
+                    print('Expected:')
+                    print(expected_output)
+                    print('Got:')
+
                 print(result)
     else:
         print(f'Test {in_name}')
@@ -56,6 +75,8 @@ for in_name, out_name in get_test_names(args['testdir']):
         print(result)
 
     print('------------------------------------------')
-print(f'Passed: {passed}')
-print(f'Failed: {failed}')
-print(f'Total: {passed + failed}')
+
+if not args['silent']:
+    print(f'Passed: {passed}')
+    print(f'Failed: {failed}')
+    print(f'Total: {passed + failed}')
